pyproject.toml
[project]
name = "gleanr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.8",
    "lxml-html-clean>=0.4.1",
    "newspaper4k>=0.9.3.1",
    "psycopg>=3.2.6",
    "psycopg-binary>=3.2.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
